[#4] WIP, linting, remove bad modules again
